add config
