+++ auto_cache_examples/example_wf.py
@@ -3,13 +3,13 @@
 
 
 from flytekit import task, workflow
+from flytekit.auto_cache.auto_cache import AutoCacheConfig
 
 
-@task
+@task(auto_cache_config=AutoCacheConfig(check_task=True, check_modules=True, check_packages=True))
 def say_hello() -> str:
     import numpy as np
     import pandas as pd
-
     from something_to_import import my_function
 
     # something dumb again
