+++ flytekit/core/task.py
@@ -1,18 +1,12 @@
 from __future__ import annotations
 
-import ast
 import datetime
-import hashlib
-import importlib
-import inspect
 import os
-import site
-import time
+from dataclasses import asdict
 from functools import update_wrapper
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Type, TypeVar, Union, overload
 
-import pkg_resources
-
+from flytekit.auto_cache.auto_cache import AutoCache, AutoCacheConfig
 from flytekit.core.utils import str2bool
 
 try:
@@ -110,6 +104,7 @@ def task(
     cache: bool = ...,
     cache_serialize: bool = ...,
     cache_version: str = ...,
+    auto_cache_config: AutoCacheConfig = ...,
     cache_ignore_input_vars: Tuple[str, ...] = ...,
     retries: int = ...,
     interruptible: Optional[bool] = ...,
@@ -148,6 +143,7 @@ def task(
     cache: bool = ...,
     cache_serialize: bool = ...,
     cache_version: str = ...,
+    auto_cache_config: AutoCacheConfig = ...,
     cache_ignore_input_vars: Tuple[str, ...] = ...,
     retries: int = ...,
     interruptible: Optional[bool] = ...,
@@ -179,172 +175,13 @@ def task(
 ) -> Union[Callable[P, FuncOut], PythonFunctionTask[T]]: ...
 
 
-# Path to the root of the current repo
-repo_root = os.path.abspath(".")
-# Get paths to all site-packages or dist-packages directories
-site_package_paths = set(site.getsitepackages() + [site.getusersitepackages()])
-
-# A cache to store already visited modules to prevent cycles
-visited_modules = set()
-
-
-def hash_ast_with_dependencies(func):
-    # Hash the AST of the main function
-    function_hash = hash_ast(func)
-
-    # Find all imported modules and their source code
-    imports = find_imports(func)
-    for imported_func in imports:
-        try:
-            # Recursively hash internal modules, hash version of external packages
-            module = inspect.getmodule(imported_func)
-            if module is not None:
-                if is_internal_module(module):
-                    module_hash = hash_module_recursive(module)
-                else:
-                    module_hash = hash_external_package_version(module)
-
-                function_hash = combine_hashes(function_hash, module_hash)
-        except (TypeError, OSError):
-            # Handle cases where the source code is not available (e.g., compiled extensions)
-            pass
-
-    return function_hash
-
-
-def is_internal_module(module):
-    """Checks if the module is part of the current repo or an external package."""
-    module_path = os.path.abspath(inspect.getfile(module))
-
-    # Check if the module is within any of the site-packages directories
-    if any(module_path.startswith(site_package_path) for site_package_path in site_package_paths):
-        return False  # It's an external package
-
-    return module_path.startswith(repo_root)
-
-
-def hash_module_recursive(module):
-    """Recursively hash the module and its dependencies if it's internal."""
-    if module.__name__ in visited_modules:
-        return ""  # Skip if already visited
-
-    visited_modules.add(module.__name__)
-
-    try:
-        # Get the source of the module
-        source = inspect.getsource(module)
-        module_hash = hashlib.sha256(source.encode("utf-8")).hexdigest()
-
-        # Find imports in the module and hash them recursively
-        module_imports = find_module_imports(module)
-        for sub_module in module_imports:
-            if is_internal_module(sub_module):
-                sub_module_hash = hash_module_recursive(sub_module)
-            else:
-                sub_module_hash = hash_external_package_version(sub_module)
-
-            module_hash = combine_hashes(module_hash, sub_module_hash)
-
-        return module_hash
-    except (TypeError, OSError):
-        return ""
-
-
-def hash_external_package_version(module):
-    """Hashes the version of an external package."""
-    try:
-        package_name = module.__name__.split(".")[0]
-        version = pkg_resources.get_distribution(package_name).version
-        package_name_version = package_name + version
-        return hashlib.sha256(package_name_version.encode("utf-8")).hexdigest()
-    except pkg_resources.DistributionNotFound:
-        return ""  # If we can't find the version (e.g., built-in libraries), return empty hash
-
-
-def combine_hashes(existing_hash, new_hash):
-    """Combine existing hash with new hash and return a new hash."""
-    combined_hash = hashlib.sha256()
-    combined_hash.update(existing_hash.encode("utf-8"))
-    combined_hash.update(new_hash.encode("utf-8"))
-    return combined_hash.hexdigest()
-
-
-def find_module_imports(module):
-    """Finds all the modules that are imported within the given module."""
-    imported_modules = []
-    try:
-        # Get the source code of the module
-        source = inspect.getsource(module)
-        tree = ast.parse(source)
-
-        class ModuleImportVisitor(ast.NodeVisitor):
-            def visit_Import(self, node):
-                for alias in node.names:
-                    imported_module = importlib.import_module(alias.name)
-                    imported_modules.append(imported_module)
-
-            def visit_ImportFrom(self, node):
-                imported_module = importlib.import_module(node.module)
-                imported_modules.append(imported_module)
-
-        visitor = ModuleImportVisitor()
-        visitor.visit(tree)
-    except Exception:
-        pass
-
-    return imported_modules
-
-
-def find_imports(func):
-    source = inspect.getsource(func)
-    tree = ast.parse(source)
-    imported_funcs = []
-
-    class ImportVisitor(ast.NodeVisitor):
-        def __init__(self):
-            self.imports = {}
-
-        def visit_Import(self, node):
-            for alias in node.names:
-                module = importlib.import_module(alias.name)
-                self.imports[alias.asname or alias.name] = module
-
-        def visit_ImportFrom(self, node):
-            module = importlib.import_module(node.module)
-            for alias in node.names:
-                obj = getattr(module, alias.name)
-                self.imports[alias.asname or alias.name] = obj
-
-        def visit_Attribute(self, node):
-            if isinstance(node.value, ast.Name) and node.value.id in self.imports:
-                obj = getattr(self.imports[node.value.id], node.attr)
-                imported_funcs.append(obj)
-
-        def visit_Call(self, node):
-            # Handle direct function calls (like some_function())
-            if isinstance(node.func, ast.Name) and node.func.id in self.imports:
-                imported_funcs.append(self.imports[node.func.id])
-            self.generic_visit(node)  # Continue visiting other nodes
-
-    visitor = ImportVisitor()
-    visitor.visit(tree)
-
-    return imported_funcs
-
-
-def hash_ast(func):
-    source = inspect.getsource(func)
-    parsed_ast = ast.parse(source)
-    ast_bytes = ast.dump(parsed_ast).encode("utf-8")
-    return hashlib.sha256(ast_bytes).hexdigest()
-
-
 def task(
     _task_function: Optional[Callable[P, FuncOut]] = None,
     task_config: Optional[T] = None,
     cache: bool = False,
     cache_serialize: bool = False,
     cache_version: str = "",
+    auto_cache_config: AutoCacheConfig = AutoCacheConfig(),
     cache_ignore_input_vars: Tuple[str, ...] = (),
     retries: int = 0,
     interruptible: Optional[bool] = None,
@@ -504,17 +341,17 @@ def task(
     """
 
     def wrapper(fn: Callable[P, Any]) -> PythonFunctionTask[T]:
-        start_time = time.time()
-        function_hash = hash_ast_with_dependencies(fn)
-        end_time = time.time()
-        elapsed_time = end_time - start_time
-        print(f"!!!!!!!!!!!!!!!!Time taken: {elapsed_time} seconds")
-        print(function_hash)
+        task_hash = ""
+        if any(asdict(auto_cache_config).values()):
+            auto_cache = AutoCache(config=auto_cache_config)
+            task_hash = auto_cache.hash_task(fn)
+
+        print("cache_version: ", cache_version if not task_hash else task_hash)
 
         _metadata = TaskMetadata(
             cache=cache,
             cache_serialize=cache_serialize,
-            cache_version=cache_version,
+            cache_version=cache_version if not task_hash else task_hash,
             cache_ignore_input_vars=cache_ignore_input_vars,
             retries=retries,
             interruptible=interruptible,
